tests/wpt/tests/css/css-box/parsing/clear-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS basic box model: parsing clear with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-box-3/#propdef-clear">
<link rel="help" href="https://drafts.csswg.org/css-logical/#float-clear">
<meta name="assert" content="clear supports the full grammar 'none | left | right | both'.">
<meta name="assert" content="clear also supports inline-start and inline-end.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("clear", "none");
test_valid_value("clear", "left");
test_valid_value("clear", "right");
test_valid_value("clear", "both");

test_valid_value("clear", "inline-start");
test_valid_value("clear", "inline-end");
</script>
</body>
</html>
